Move MomentAccumulator from moments.py into util.py
